Revamp Add_invoice UI and client-side logic
Major refactor of Add_invoice.aspx: redesigned UI/CSS, modernized form layout (grid-based panels), integrated Select2 and updated jQuery/datepickers, and consolidated client-side validation and realtime calculations (quantity/rate/discount/tax/net). Added protections: over-invoicing snap-back, zero-total submission prevention, improved footer recalculation, and UX/interaction enhancements (sticky headers, responsive grids, buttons). Other changes: switched Web.config DbConn to UAT, added SystemErrors log entry, appended audit log for admin password update, and appended invoice generation entry to InvoiceLogs. Several code-behind and designer files for Add_invoice, Manual_Invoice, View_Invoice, seartch_invoice and print pages were updated to align with the new UI/logic.

+++ Bill_Software/corporate/business/print/NewInvoiceDuplicate.aspx.cs
@@ -31,7 +31,7 @@ namespace Bill_Software.corporate.business.print
         public string taxorvat = "";
         public string proOrser = "";
         public string psid = "";
-        public int TQ = 0;
+        public double TQ = 0;
         public string lblSubtotal = "", lbldiscount = "", lblstax = "", lblstax0 = "", lblnetamount = "", lblword = "";
 
         // New variables for Freight/Other
@@ -349,7 +349,7 @@ namespace Bill_Software.corporate.business.print
                     {
                         string HSN = dtp.Rows[i]["HSN"].ToString();
                         string Productname = dtp.Rows[i]["Product_name"].ToString();
-                        int Quantity = Convert.ToInt32(dtp.Rows[i]["Quantity"]);
+                        double Quantity = Convert.ToDouble(dtp.Rows[i]["Quantity"]);
                         TQ += Quantity;
 
                         double baserate = Math.Round(Convert.ToDouble(dtp.Rows[i]["sail_rate"]), 2);
